+++ golang/kurtosis-lambda/impl/example_kurtosis_lambda_args.go
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2021 - present Kurtosis Technologies LLC.
+ * All Rights Reserved.
+ */
+
+package impl
+
+type ExampleKurtosisLambdaArgs struct {
+	// Indicates the log level for this Kurtosis Lambda implementation
+	LogLevel string `json:"logLevel"`
+}