+++ golang/kurtosis-lambda/main.go
@@ -1,0 +1,35 @@
+/*
+ * Copyright (c) 2021 - present Kurtosis Technologies LLC.
+ * All Rights Reserved.
+ */
+
+package main
+
+import (
+	"fmt"
+	"github.com/kurtosis-tech/kurtosis-lambda-api-lib/golang/lib/execution"
+	"github.com/kurtosis-tech/kurtosis-lambda-starter-pack/golang/kurtosis-lambda/impl"
+	"github.com/sirupsen/logrus"
+	"os"
+)
+
+const (
+	successExitCode = 0
+	failureExitCode = 1
+)
+
+func main() {
+
+	// >>>>>>>>>>>>>>>>>>> REPLACE WITH YOUR OWN CONFIGURATOR <<<<<<<<<<<<<<<<<<<<<<<<
+	configurator := impl.NewExampleKurtosisLambdaConfigurator()
+	// >>>>>>>>>>>>>>>>>>> REPLACE WITH YOUR OWN CONFIGURATOR <<<<<<<<<<<<<<<<<<<<<<<<
+
+
+	lambdaExecutor := execution.NewKurtosisLambdaExecutor(configurator)
+	if err := lambdaExecutor.Run(); err != nil {
+		logrus.Errorf("An error occurred running the kurtosis-lambda executor:")
+		fmt.Fprintln(logrus.StandardLogger().Out, err)
+		os.Exit(failureExitCode)
+	}
+	os.Exit(successExitCode)
+}