+++ docs/changelog.md
@@ -1,4 +1,18 @@
 # TBD
 
+# 0.1.1
+### Features
+* Added `ExampleLambda` which is an example of how to implement a Kurtosis Lambda
+* Added `ExampleLambdaConfigurator` and `ExampleLambdaArgs` which are used to start the `ExampleLambda`implementation 
+* Added the main go file for startup the Lambda
+* Added Dockerfile with build and run configurations
+* Added Circle CI configuration file
+* Added golang directory and go mod file
+* Added shell script to execute kurtosis release repository script for this repository
+* Implement `ExampleLambda` execute method
+* Added bootstrap script; users can use it to start creating their own implementation of Kurtosis Lambda
+* Added supported-languages file
+* Added script for validating bootstrap during CI
+
 #0.1.0
 * Init commit